# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ct-project-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://ct-project.pp.ua/api/"
        },
        {
            "url": "http://172.19.0.1/api/"
        }
    ],
    "paths": {
        "/v1/admin/categories": {
            "get": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "3133d5ea8706f58a99df9c5bdbede62b",
                "parameters": [
                    {
                        "name": "lastId",
                        "in": "query",
                        "description": "Min: 0",
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "7a001a43920c3ba28c4b3df77ae1311b",
                "parameters": [
                    {
                        "name": "parentId",
                        "in": "query",
                        "description": "Parent Category ID",
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Name of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Show created category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/admin/categories/{id}": {
            "get": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "2d471267bdc73dcc22ac5a4447d2800d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show city by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "0a74440c8eb0366b196f573bb5038c4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The category has been deleted"
                    },
                    "400": {
                        "description": "The category used by other services.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "message": "This category is using by other services.",
                                        "code": 400
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "ee196a96e4c2e31deca15254b8f59d64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parentId",
                        "in": "query",
                        "description": "Parent Category ID",
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Name of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all cities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "description": "Show information about category",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "parentId": {
                        "description": "Integer or null",
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Support"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Support"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": [
                    {
                        "id": 1,
                        "parentId": null,
                        "title": "Макіяж",
                        "slug": "makiyazh",
                        "createdBy": {
                            "id": 1,
                            "firstName": "Марк",
                            "lastName": "Косенко",
                            "email": "mark.kosenko@gmail.com"
                        },
                        "updatedBy": {
                            "id": 6,
                            "firstName": "Ганна",
                            "lastName": "Носенко",
                            "email": "hanna.nosenko@gmail.com"
                        },
                        "createdAt": "2023-12-02 14:51:18",
                        "updatedAt": "2023-12-12 16:26:18"
                    }
                ]
            },
...
            "ValidationErrors": {
                "description": "Validation Errors",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "description": "each key describes error message",
                        "properties": {
                            "query or path name": {
                                "type": "array",
                                "items": {}
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The email field must be a valid email address. (and 3 more errors)",
                    "errors": {
                        "email": [
                            "The email field must be a valid email address."
                        ],
                        "password": [
                            "The password field confirmation does not match.",
                            "The password field must contain at least one uppercase and one lowercase letter.",
                            "The password field must contain at least one number."
                        ]
                    }
                }
            },
...
            "Support": {
                "description": "Show information about user with support role",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}